Decomposed integration and performance tests
