+++ .github/workflows/tests.yml
@@ -77,3 +77,8 @@ jobs:
       - name: Notebook tests
         run: poetry run pytest --no-cov --nbval-lax -p no:python src
         shell: bash
+
+      - name: Performance tests
+        if: runner.os == 'Linux'
+        run: poetry run pytest --no-cov tests/integration_tests
+        shell: bash
