+++ tests/performance_tests/oad_process/data/oad_process.toml
@@ -1,0 +1,71 @@
+title = "Sample OAD Process"
+
+# List of folder paths where user added custom registered OpenMDAO components
+module_folders = []
+
+# Input and output files
+input_file = "../results/problem_inputs.xml"
+output_file = "../results/problem_outputs.xml"
+
+# Definition of problem driver assuming the OpenMDAO convention "import openmdao.api as om"
+driver = "om.ScipyOptimizeDriver(tol=1e-6, optimizer='SLSQP')"
+
+# Definition of OpenMDAO model
+[model]
+    # Solvers are defined assuming the OpenMDAO convention "import openmdao.api as om"
+    nonlinear_solver = "om.NonlinearBlockGS(maxiter=100, iprint=1)"
+    linear_solver = "om.DirectSolver()"
+    # Although "model" is a mandatory name for the top level of the model, its
+    # sub-components can be freely named by user
+    [model.wing_position_loop]
+        nonlinear_solver = "om.BroydenSolver(maxiter=20, iprint=0)"
+        linear_solver = "om.DirectSolver()"
+        [model.wing_position_loop.cg_tail_loops]
+            nonlinear_solver = "om.NonlinearBlockGS(maxiter=50, iprint=0)"
+            linear_solver = "om.DirectSolver()"
+            [model.wing_position_loop.cg_tail_loops.geometry]
+                # An OpenMDAO component is identified by its "id"
+                id = "fastoad.geometry.legacy"
+            [model.wing_position_loop.cg_tail_loops.weight]
+                id = "fastoad.weight.legacy"
+            [model.wing_position_loop.cg_tail_loops.mtow]
+                id = "fastoad.loop.mtow"
+            [model.wing_position_loop.cg_tail_loops.hq_tail_sizing]
+                id = "fastoad.handling_qualities.tail_sizing"
+
+        [model.wing_position_loop.hq_static_margin]
+            id = "fastoad.handling_qualities.static_margin"
+        [model.wing_position_loop.wing_position]
+            id = "fastoad.loop.wing_position"
+
+    [model.aerodynamics_highspeed]
+        id = "fastoad.aerodynamics.highspeed.legacy"
+    [model.aerodynamics_lowspeed]
+        id = "fastoad.aerodynamics.lowspeed.legacy"
+    [model.aerodynamics_takeoff]
+        id = "fastoad.aerodynamics.takeoff.legacy"
+    [model.aerodynamics_landing]
+        id = "fastoad.aerodynamics.landing.legacy"
+        use_xfoil = false
+    [model.performance]
+        id = "fastoad.performances.breguet"
+        propulsion_id = "fastoad.wrapper.propulsion.rubber_engine"
+#    [model.performance]
+#        id = "fastoad.performances.sizing_mission"
+#        propulsion_id = "fastoad.wrapper.propulsion.rubber_engine"
+#        out_file = "./flight_points.csv"
+    [model.wing_area]
+        id = "fastoad.loop.wing_area"
+
+[optimization]  # This section is needed only if optimization process is run
+    [[optimization.design_var]]
+        name = "data:geometry:wing:aspect_ratio"
+        lower = 9.0
+        upper = 12.0
+
+    [[optimization.constraint]]
+        name = "data:geometry:wing:span"
+        upper = 36.0
+
+    [[optimization.objective]]
+        name = "data:mission:sizing:fuel"
